iterate on suggestion and move interfaces
