+++ src/token.ts
@@ -9,3 +9,29 @@ export interface IJupybaseClient {
 export const IJupybaseClient = new Token<IJupybaseClient>(
   'jupybase:IJupybaseClient'
 );
+
+export interface IEnvironmentContent {
+  kernelEnv: string;
+  buildEnv: string;
+  dependencies: string[];
+  lockfile: string;
+}
+
+/**
+ * A promise that resolves to project and environment information.
+ */
+export interface IProjectInfo {
+  factory: () => Promise<{
+    name: string;
+    description: string;
+    details: string;
+    environment: IEnvironmentContent;
+  }>;
+}
+
+/**
+ * A token for the plugin that provides project information.
+ */
+export const IProjectInfo = new Token<IProjectInfo>(
+  'jupybase:project-info-provider'
+);
